Move gallery examples module into lexical-website package
The gallery example definitions are website content, so they belong in
the website package. The screenshot script now imports from there.
Removes the @gallery-examples webpack alias since pluginList.tsx can
use a simple relative import.
